Fixes behavior of shutdown app on stop in SimulationContext (#254)
This MR fixes the behavior of
`omni.isaac.orbit.sim.SimulationCfg.shutdown_app_on_stop` to not close
the app when the flag is False. Earlier, it would throw an error
(invalid physics handles) when the stop button was pressed in the GUI
and the flag was False.

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there