Fixes invisible to secondary rays for markers (#316)
# Description

Earlier, the markers set "invisible" to secondary rays on the
wrong prim (XForm instead of mesh). For context, this property is used
when we don't want to see the marker prims on secondary rendering
operations used to compute depth and semantic images. However,
the markers are still visible on RGB.

This MR makes sure that the marker prims are invisible to these rays by
setting the property on the correct prim.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Screenshots

| Before | After |
| ------ | ----- |
|
![earlier](https://github.com/isaac-orbit/orbit/assets/12863862/140f0afb-86bc-48e9-9812-7e7d9489cecb)
|
![now](https://github.com/isaac-orbit/orbit/assets/12863862/643b1ed3-b454-4a94-b2e7-256aed012d13)
|

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there