Adds ability to launch multiple jobs on cluster with different code changes (#832)
# Description

Currently, for multiple runs, the code is pushed to the same location on
the cluster. When the runs are not directly started (i.e.,
are in the queue), the code will be replaced each time a new job is
submitted. This PR creates a new copy of the code for each job to ensure
that the intended version is used for the submitted job.

## Type of change

- New feature (non-breaking change which adds functionality)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have run all the tests with `./isaaclab.sh --test` and they pass
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there